requirements.txt
fastapi
uvicorn
pypdf
python-dotenv
openai
python-multipart
